move file collections under input
