+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@yuhuinc/yuhui",
-  "version": "0.1.10-rc.7",
+  "version": "0.1.12",
   "main": "dist/index.js",
   "scripts": {
     "build": "rm -rf dist/ && tsc",
