+++ stories/components/Input/InputCollection.md
