+++ stories/components/Input/InputNumber.md
